third_party/WebKit/LayoutTests/editing/inserting/insert_div_text_into_text.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            'There b|e no line breaks in this sentence.',
        '</div>',
    ].join(''),
    'insertHTML <div>XYZ</div>',
    [
        '<div contenteditable>',
            'There bXYZ|e no line breaks in this sentence.',
        '</div>',
    ].join('')),
    'insertHTML with DIV to text');
</script>
